requirements.txt
requests==2.31.0
rdflib==7.0.0
sparqlwrapper==2.0.0
openai>=1.6.1
langchain==0.0.350
langchain-openai>=0.0.2
chromadb==0.4.18
sentence-transformers==2.2.2
python-dotenv==1.0.0
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
